+++ tests/test_pawn.py
@@ -1,6 +1,22 @@
 import unittest
 from piezas.pawn import Pawn
 from game.board import Board
+from piezas.piece import Piece
+
+
+class MockBoard:
+    def __init__(self):
+        self.board = [[None for _ in range(8)] for _ in range(8)]
+
+    def get_piece(self, row, col):
+        return self.board[row][col]
+
+    def set_piece(self, row, col, piece):
+        self.board[row][col] = piece
+
+class MockPiece:
+    def __init__(self, color):
+        self.color = color
 
 class TestPawn(unittest.TestCase):
 
@@ -30,5 +46,22 @@ class TestPawn(unittest.TestCase):
         self.assertEqual(resultados, movimientos) 
 
     
+    def test_pawn_captures_diagonally(self):
+        # Crear el tablero vacío
+        board = MockBoard()
+        # Añadir una pieza enemiga en diagonal para que el peón pueda capturarla
+        enemy_piece = MockPiece('black')  #La pieza enemiga es negra
+        board.set_piece(3, 2, enemy_piece)  #Colocar la pieza en (3, 2)
+        # Crear el peón blanco en la posición (2, 1)
+        pawn = Pawn("white") 
+        posibles = pawn.movimiento_hacia_adelante(2, 1, False, 'white', board)
+        # Verificar que el peón puede capturar la pieza en (3, 2)
+        self.assertIn((3, 2), posibles)
+
+
+
+
+if __name__ == '__main__':
+    unittest.main()